backend/iNOteAPI/requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.3
firebase-admin==2.13.0
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-cloud-texttospeech==0.2.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.0
idna==2.7
msgpack==0.5.6
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
toml==0.10.0
urllib3==1.24.1
